backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
openai
requests
python-multipart
